+++ WP-PHP7.4/woo-scalexpert/Views/simulation/product.php
@@ -14,9 +14,9 @@
 			$solutions        = $simulations;
 			$md5GroupSolution = md5( 'all' . sizeof( $solutions ) . rand( 0, 100 ) );
 			$isSimulation     = "product";
-			
-			include( plugin_dir_path( __FILE__ ) . 'solution.php' );
-			include( plugin_dir_path( __FILE__ ) . 'modalSimulation.php' );
+
+            include_once( plugin_dir_path( __FILE__ ) . 'solution.php' );
+            include_once( plugin_dir_path( __FILE__ ) . 'modalSimulation.php' );
 		}
 	?>
 </div>