+++ WP-PHP8.2/woo-scalexpert/Views/cartFinancialSimulationContent.php
@@ -1,0 +1,18 @@
+<?php
+    /**
+     * Copyright © Scalexpert.
+     * This file is part of Scalexpert plugin for Wordpress / Woocommerce.
+     *
+     * @author    Société Générale
+     * @copyright Scalexpert
+     */
+?>
+<div class="sep-Simulations">
+    <?php
+        
+        
+        include_once( plugin_dir_path( __FILE__ ) . 'simulation/cart.php' );
+    
+    ?>
+</div>
+